teonitesp/requirements.txt
asn1crypto==0.24.0
attrs==17.4.0
Automat==0.6.0
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.1.4
cssselect==1.0.3
Django==2.0.2
django-rest-framework==0.1.0
djangorestframework==3.7.7
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
lxml==4.1.1
Markdown==2.6.11
parsel==1.4.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==17.5.0
pytz==2018.3
queuelib==1.4.2
requests==2.18.4
Scrapy==1.5.0
scrapy-djangoitem==1.1.1
service-identity==17.0.0
six==1.11.0
Twisted==17.9.0
urllib3==1.22
w3lib==1.19.0
zope.interface==4.4.3
